aider: Made the ChatMessages scrollable by updating CSS classes.
# Aider chat conversation:

USER: The ChatMessages are not scrollable

ASSISTANT: To make the ChatMessages scrollable, we need to ensure that the container has the appropriate CSS classes to enable scrolling. We will update the `ChatMessages.svelte` file to add the necessary classes.

Here is the *SEARCH/REPLACE* block:

src\lib\apps\ChatApp\ChatMessages.svelte
```svelte
<<<<<<< SEARCH
<div class="flex flex-col space-y-2 p-2 bg-gray-900 text-white h-full overflow-hidden" id="chatui-chat">
=======
<div class="flex flex-col space-y-2 p-2 bg-gray-900 text-white h-full overflow-y-auto" id="chatui-chat">
>>>>>>> REPLACE
```

This change will make the `ChatMessages` component scrollable.