aider: Add padding to the bottom of the `ChatMessages` component.
# Aider chat conversation:

USER: Can you add a padding to the bottom of the ChatMessages?

ASSISTANT: To add padding to the bottom of the `ChatMessages` component, we need to update the `ChatMessages.svelte` file.

Here is the *SEARCH/REPLACE* block to achieve this:

src\lib\apps\ChatApp\ChatMessages.svelte
```svelte
<<<<<<< SEARCH
<div class="flex flex-col space-y-2 p-2 bg-gray-900 text-white h-full overflow-y-auto scrollbar-thin scrollbar-thumb-gray-700 scrollbar-track-gray-900" id="chatui-chat" bind:this={chatContainer}>
=======
<div class="flex flex-col space-y-2 p-2 pb-16 bg-gray-900 text-white h-full overflow-y-auto scrollbar-thin scrollbar-thumb-gray-700 scrollbar-track-gray-900" id="chatui-chat" bind:this={chatContainer}>
>>>>>>> REPLACE
```

This change will add padding to the bottom of the `ChatMessages` component.